Refactor some things
