+++ app/crud/async_db/vector_tiles/max_date.py
@@ -1,11 +1,10 @@
 from ....application import db
 from ....errors import RecordNotFoundError
-from .nasa_viirs_fire_alerts import SCHEMA
 
 
-async def get_max_date(version):
+async def get_max_date(version, schema):
     t = db.table(version)
-    t.schema = SCHEMA
+    t.schema = schema
     sql = db.select(
         [db.literal_column("max(alert__date)").label("Max alert date")]
     ).select_from(t)
