+++ app/routes/nasa_viirs_fire_alerts/esri_vector_tile_server.py
@@ -2,8 +2,9 @@ from typing import Any, Dict
 
 from fastapi import APIRouter, Depends, Path, Response
 
+from . import default_start, default_end
 from ...models.pydantic.esri import VectorTileService
-from ...routes import DATE_REGEX, default_end, default_start
+from ...routes import DATE_REGEX
 from ...routes.esri_vector_tile_server import _get_vector_tile_server
 from ...routes.nasa_viirs_fire_alerts.vector_tiles import (
     dataset,
