+++ app/routes/umd_modis_burned_areas/__init__.py
@@ -1,5 +1,6 @@
 from typing import List, Optional, Type
 
+import pendulum
 from aenum import extend_enum
 from fastapi import Query
 
@@ -16,28 +17,13 @@ from ...models.enumerators.tile_caches import TileCacheType
 latest_version: Optional[str] = get_latest_version(
     dataset_name, TileCacheType.dynamic_vector_tile_cache
 )
-if latest_version:
-    included_attribute_type: Type[Attributes] = get_attributes_enum(latest_version)
-else:
-    included_attribute_type = Attributes
-    for field in default_attributes:
-        extend_enum(included_attribute_type, field, field)
 
 
-IncludedAttributes = Optional[List[included_attribute_type]]  # type: ignore
+def default_start():
+    now = pendulum.now()
+    return now.subtract(months=1).to_date_string()
 
 
-async def include_attributes(
-    include_attribute: IncludedAttributes = Query(  # type: ignore
-        default_attributes,
-        title="Included Attributes",
-        description="Select which attributes to include in vector tile. Will always show attribute count. "
-        "Documentation list available attributes of latest version. For legacy version "
-        "please check data-api for available attribute values.",
-    ),
-) -> List[str]:
-    attributes: List[str] = list()
-    if include_attribute:
-        for attribute in include_attribute:
-            attributes.append(attribute.value)  # type: ignore
-    return attributes
+def default_end():
+    now = pendulum.now()
+    return now.to_date_string()
